added more support for multiframe packages. bumped blib version number to 1.1.8 bumped blinkentools version number to 2.10
git-svn-id: file:///home/daniel/src/blinkenlights/legacy/svn/blinkenlights/trunk@348 fe0de55c-ccef-0310-be5c-e8635c8c2365

+++ code/blib/blib/bpacket.h
@@ -44,7 +44,7 @@ BPacket * b_packet_new  (gint     width,
 			 gint     channels,
 			 gint     maxval,
 			 guint    magic,
-			 gint    *data_size);
+			 gsize   *data_size);
 gsize     b_packet_size (BPacket *packet);
 void      b_packet_hton (BPacket *packet);
 void      b_packet_ntoh (BPacket *packet);
